+++ src/Components/Audio/Audio.js
@@ -3,30 +3,30 @@ import AudioPlayer from 'react-h5-audio-player'
 import 'react-h5-audio-player/lib/styles.css'
 import React, {useState, useEffect} from 'react'
 
-const Audio = () => {
+const Audio = (props) => {
 
-const [audio, setAudio] = useState([])
+// const [audio, setAudio] = useState([])
 
-useEffect(()=>{
+// useEffect(()=>{
  
-axios.get(`/api/track/`)
-.then((res)=>{
-    setAudio(res.data)
-})
-.catch(err=>console.log(err))
+// axios.get(`/api/track/`)
+// .then((res)=>{
+//     setAudio(res.data)
+// })
+// .catch(err=>console.log(err))
 
-},[])
-    const mappedTrack = audio.map(ele =>ele.track)
-console.log(mappedTrack)
-    console.log(audio[0])
+// },[])
+//     const mappedTrack = audio.map(ele =>ele.track)
+// console.log(mappedTrack)
+//     console.log(audio[0])
 
     return(
 
         <div>
-        <span>Audio Player</span>
+        <span><h2>Fresh Tracks</h2></span>
     <AudioPlayer
       autoPlay
-      src={mappedTrack}
+      src={props.mappedTrack}
       onPlay={e => console.log("onPlay")}
       // other props here
     />
